requirements.txt
python-dotenv
requests
twitchio
streamlink
tqdm
yt-dlp
